Add slow xslt that performs the merge
The merge _can_ be done just using XSLT, but it appears that it's O(n^2)
complexity, so only works for small documents.

I don't know whether there's a way to tell the transform to index values
of interest, or whether the problem is just that xsltproc doesn't do
this optimally.