XSLT to merge heart rate TCX into workout TCX
Now have an XSLT-only solution to merging the TCX files. Use `key()` to
create an index, and `<xsl:for-each` to ensure the proper context when
retrieving data from the index.

Was able to process a 16,000+ point workout file with slightly larger
heart rate TCX file in less than 0.3 seconds.

Deleted the bash/xslt/awk script.